wine-quality/requirements.txt
numpy
pandas
scikit-learn>=1.2
mlflow>=2.10
pyyaml
joblib
matplotlib
streamlit>=1.33
dvc>=3.0